# docs/openapi/pdt-openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Platform Data Toolkit (PDT) API",
        "description": "PDT is a centralized API server that serves as an enterprise knowledge silo for company-specific concepts, documents, and their relationships.\n\n## Features\n\n- **Knowledge Asset Management**: Store and manage knowledge assets with rich metadata\n- **Tag-Based Classification**: Multi-dimensional tagging including asset type, business domain, language, etc.\n- **Graph-Based Relationships**: Define and traverse relationships between assets\n- **Concept Collections**: Organize assets into named collections\n- **Search & Discovery**: Full-text search and tag-based filtering\n- **Audit Logging**: Track all changes with user attribution\n- **Authorization Context**: Cedar-based visibility, owner groups, and confidentiality management",
        "contact": {
            "name": "PDT Team"
        },
        "license": {
            "name": "MIT OR Apache-2.0"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/assets/{id}/history": {
            "get": {
                "tags": [
                    "audit"
                ],
                "summary": "Get history for a specific asset",
                "operationId": "get_asset_history",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Asset ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Asset change history",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AuditEntry"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Asset not found"
                    }
                }
            }
        },
...
        "/api/audit": {
            "get": {
                "tags": [
                    "audit"
                ],
                "summary": "List audit entries",
                "operationId": "list_audit_entries",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "entity_type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "entity_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of audit entries",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedResponse_AuditEntry"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/collections": {
            "get": {
                "tags": [
                    "collections"
                ],
                "summary": "List collections with pagination",
                "operationId": "list_collections",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of collections",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PaginatedResponse_Collection"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "collections"
                ],
                "summary": "Create a new collection",
                "operationId": "create_collection",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCollectionRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Collection created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Collection"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/collections/{id}": {
            "get": {
                "tags": [
                    "collections"
                ],
                "summary": "Get collection by ID",
                "operationId": "get_collection",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Collection ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Collection found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Collection"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Collection not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "collections"
                ],
                "summary": "Update a collection",
                "operationId": "update_collection",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Collection ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCollectionRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Collection updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Collection"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Collection not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "collections"
                ],
                "summary": "Delete a collection",
                "operationId": "delete_collection",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Collection ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Collection deleted successfully"
                    },
                    "404": {
                        "description": "Collection not found"
                    }
                }
            }
        },
        "/api/collections/{id}/assets": {
            "post": {
                "tags": [
                    "collections"
                ],
                "summary": "Add an asset to a collection",
                "operationId": "add_asset",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Collection ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddAssetRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Asset added to collection successfully"
                    },
                    "404": {
                        "description": "Collection or asset not found"
                    }
                }
            }
        },
        "/api/collections/{id}/assets/{asset_id}": {
            "delete": {
                "tags": [
                    "collections"
                ],
                "summary": "Remove an asset from a collection",
                "operationId": "remove_asset",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Collection ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "asset_id",
                        "in": "path",
                        "description": "Asset ID to remove",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Asset removed from collection successfully"
                    },
                    "404": {
                        "description": "Collection or asset not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AddAssetRequest": {
                "type": "object",
                "description": "Request to add an asset to a collection",
                "required": [
                    "asset_id"
                ],
                "properties": {
                    "asset_id": {
                        "type": "string"
                    }
                }
            },
            "AddTagRequest": {
                "type": "object",
                "description": "Request to add a tag to an asset",
                "required": [
                    "category",
                    "value"
                ],
                "properties": {
                    "category": {
                        "type": "string"
                    },
                    "value": {
                        "type": "string"
                    }
                }
            },
...
            "AuditAction": {
                "type": "string",
                "description": "Types of audited actions",
                "enum": [
                    "create",
                    "update",
                    "delete",
                    "add_tag",
                    "remove_tag",
                    "add_relation",
                    "remove_relation",
                    "add_to_collection",
                    "remove_from_collection"
                ]
            },
            "AuditEntry": {
                "type": "object",
                "description": "Audit log entry",
                "required": [
                    "_id",
                    "entity_type",
                    "entity_id",
                    "action",
                    "changes",
                    "user_id",
                    "timestamp"
                ],
                "properties": {
                    "_id": {
                        "type": "string"
                    },
                    "action": {
                        "$ref": "#/components/schemas/AuditAction"
                    },
                    "changes": {},
                    "entity_id": {
                        "type": "string"
                    },
                    "entity_type": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user_id": {
                        "type": "string"
                    }
                }
            },
...
            "Collection": {
                "type": "object",
                "description": "Named collection of assets",
                "required": [
                    "_id",
                    "name",
                    "created_at",
                    "updated_at",
                    "created_by",
                    "updated_by"
                ],
                "properties": {
                    "_id": {
                        "type": "string"
                    },
                    "asset_ids": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_by": {
                        "type": "string"
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "name": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tag"
                        }
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_by": {
                        "type": "string"
                    }
                }
            },
...
            "CreateCollectionRequest": {
                "type": "object",
                "description": "Request to create a new collection",
                "required": [
                    "name"
                ],
                "properties": {
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "name": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AddTagRequest"
                        }
                    }
                }
            },
...
            "PaginatedResponse_AuditEntry": {
                "type": "object",
                "description": "Paginated response wrapper",
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "description": "Audit log entry",
                            "required": [
                                "_id",
                                "entity_type",
                                "entity_id",
                                "action",
                                "changes",
                                "user_id",
                                "timestamp"
                            ],
                            "properties": {
                                "_id": {
                                    "type": "string"
                                },
                                "action": {
                                    "$ref": "#/components/schemas/AuditAction"
                                },
                                "changes": {},
                                "entity_id": {
                                    "type": "string"
                                },
                                "entity_type": {
                                    "type": "string"
                                },
                                "timestamp": {
                                    "type": "string",
                                    "format": "date-time"
                                },
                                "user_id": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "next_cursor": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "total": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int64",
                        "minimum": 0
                    }
                }
            },
            "PaginatedResponse_Collection": {
                "type": "object",
                "description": "Paginated response wrapper",
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "description": "Named collection of assets",
                            "required": [
                                "_id",
                                "name",
                                "created_at",
                                "updated_at",
                                "created_by",
                                "updated_by"
                            ],
                            "properties": {
                                "_id": {
                                    "type": "string"
                                },
                                "asset_ids": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                },
                                "created_at": {
                                    "type": "string",
                                    "format": "date-time"
                                },
                                "created_by": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": [
                                        "string",
                                        "null"
                                    ]
                                },
                                "name": {
                                    "type": "string"
                                },
                                "tags": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Tag"
                                    }
                                },
                                "updated_at": {
                                    "type": "string",
                                    "format": "date-time"
                                },
                                "updated_by": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "next_cursor": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "total": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int64",
                        "minimum": 0
                    }
                }
            },
...
            "Tag": {
                "type": "object",
                "description": "Tag attached to an asset\n\nUses industry-standard key-value format:\n- `category`: Free-form string (e.g., \"type\", \"status\", \"domain\", \"priority\")\n- `value`: The tag value (e.g., \"document\", \"draft\", \"backend\", \"high\")\n\nCommon categories (suggestions, not enforced):\n- `type` - Asset type (document, idea, requirement, bug, task)\n- `status` - Status (draft, review, approved, archived)\n- `domain` - Business domain (backend, frontend, infra)\n- `lang` - Language (en, fa)\n- `priority` - Priority (high, medium, low)\n- `scope` - Feature/component scope",
                "required": [
                    "id",
                    "category",
                    "value",
                    "added_by",
                    "added_at"
                ],
                "properties": {
                    "added_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "added_by": {
                        "type": "string"
                    },
                    "category": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "value": {
                        "type": "string"
                    }
                }
            },
...
            "UpdateCollectionRequest": {
                "type": "object",
                "description": "Request to update a collection",
                "properties": {
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                }
            }
        }
    },
...
}